Propius/propius/controller/requirements.txt
numpy==1.21.1
redis==5.0.0
pandas==1.3.1
grpcio==1.54.3
grpcio-tools==1.54.2
pyyaml==5.3.1
matplotlib==3.4.3